+++ src/main/java/name/divinityunbound/screen/ModScreenHandlers.java
@@ -70,6 +70,10 @@ public class ModScreenHandlers {
             Registry.register(Registries.SCREEN_HANDLER, new Identifier(DivinityUnbound.MOD_ID, "proteus_converter"),
                     new ExtendedScreenHandlerType<>(ProteusConverterScreenHandler::new));
 
+    public static final ScreenHandlerType<ZeusBatteryScreenHandler> ZEUS_BATTERY_SCREEN_HANDLER =
+            Registry.register(Registries.SCREEN_HANDLER, new Identifier(DivinityUnbound.MOD_ID, "zeus_battery"),
+                    new ExtendedScreenHandlerType<>(ZeusBatteryScreenHandler::new));
+
     public static final ScreenHandlerType<SpaceTimeFurnaceScreenHandler> SPACE_TIME_FURNACE_SCREEN_HANDLER =
             Registry.register(Registries.SCREEN_HANDLER, new Identifier(DivinityUnbound.MOD_ID, "space_time_furnace"),
                     new ScreenHandlerType<>(SpaceTimeFurnaceScreenHandler::new, FeatureFlags.VANILLA_FEATURES));